+++ src/v1/Tasks/AbstractTask.js
@@ -6,14 +6,16 @@ module.exports = class AbstractTask {
     this.name = name;
   }
 
-  // asynchroneous execution within a workflow
-  dispatch() {
-    return new Engine().dispatch([this])[0];
+  // asynchronous execution within a workflow
+  async dispatch() {
+    const result = await new Engine().dispatch([this]);
+    return result[0].then(() => undefined);
   }
 
-  // synchroneous execution within a workflow
-  execute() {
-    return new Engine().execute([this])[0];
+  // synchronous execution within a workflow
+  async execute() {
+    const result = await new Engine().execute([this]);
+    return result[0];
   }
 
   static methods() {